refactor ♻️: Refactor formatters.py to core/formatters.py and update imports
This commit refactors the `formatters.py` file from the webui directory to the core directory, renaming it to `formatters.py`. The changes include updating all import statements to reflect this new location. This refactoring is part of a larger effort to standardize the project's directory structure and improve code organization.
