+++ insanely_fast_whisper_api/webui/merge_handler.py
@@ -11,7 +11,7 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple
 
-from insanely_fast_whisper_api.webui.formatters import FORMATTERS
+from insanely_fast_whisper_api.core.formatters import FORMATTERS
 
 # Configure logger
 logger = logging.getLogger("insanely_fast_whisper_api.webui.merge_handler")
